Stop logging if logger is off and added GitHub Actions CI/CD
This commit introduces .gitignore file, updated project configuration files in the .idea directory and added closing logging if logger is offline/can't be connected to in InverterData.py. A new CI/CD workflow for building and pushing Docker images using GitHub actions is also added. This is intended to automate repetitive building and deploying tasks, improve production pushes and simplify the overall process.