Add debugging feature in the InverterData script
A debugging feature has been added to the InverterData.py script. Config.cfg now includes a 'debug' configuration option. If the debug option is set to 1, verbose output and extra logging for data sent and received through the clientSocket, as well as loop operation, will become visible. This feature will facilitate the process of identifying issues and debugging.